moved EnclosureInfo class to seperate file
